E-mail változtatás esetén, előző postacím letárolásra kerül. Ha 1 napig nem aktiválja akkor visszaállításra. A levelek tartalma is változott.

+++ cron.php
@@ -11,7 +11,16 @@ foreach($users as $profile)
 		$user->update($profile);
 	}
 $users=$user->get("`status` = '1'");
-foreach($users as $profile)
-	if(isset($profile["data"]["acst"]) and strtotime($profile["data"]["acst"].' + '.$config["activation/delete"].' day')<=strtotime(date("Y-m-d h:i:s")))
+foreach($users as $profile) {
+	if(!isset($profile["data"]["old_mail"]) and isset($profile["data"]["acst"]) and strtotime($profile["data"]["acst"].' + '.$config["activation/delete"].' day')<=strtotime(date("Y-m-d h:i:s")))
 		$user->delete($profile);
+	if(isset($profile["data"]["old_mail"]) and isset($profile["data"]["acst"]) and strtotime($profile["data"]["acst"].' + '.$config["newmail/backup"].' day')<=strtotime(date("Y-m-d h:i:s"))) {
+		unset($profile["data"]["acst"]);
+		unset($profile["data"]["activation_code"]);
+		$profile["mail"] = $profile["data"]["old_mail"];
+		unset($profile["data"]["old_mail"]);
+		$profile["status"] = 2;
+		$user->update($profile);
+	}
+}
 ?>
